move main layout to imports
