+++ client/head.html
@@ -1,0 +1,9 @@
+<head>
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<noscript>
+		<div class="container-sm noscript">
+			<img src="/logo/openki.png" alt="Openki logo">
+			<h2>Please enable javascript</h2>
+		</div>
+	</noscript>
+</head>
