+++ imports/ui/layouts/app-body.html
@@ -1,13 +1,3 @@
-<head>
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<noscript>
-		<div class="container-sm noscript">
-			<img src="/logo/openki.png" alt="Openki logo">
-			<h2>Please enable javascript</h2>
-		</div>
-	</noscript>
-</head>
-
 <template name="layout">
 	{{#if showRegionSplash}}
 		{{> regionsSplash}}
