--- imports/ui/layouts/frame.js
@@ -1,3 +1,0 @@
-import '/imports/ui/components/messages/messages.js';
-
-import './frame.html';
